With_OpenAI_API_KEY/resume_optimizer_backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
openai>=1.27.0
PyPDF2>=3.0.0
fpdf2>=2.7.0
supabase>=2.4.0
python-dotenv>=1.0.0
pydantic>=2.7.0
python-multipart>=0.0.6
